fix bug

+++ src/main/java/com/dancecube/image/UserRatioImage.java
@@ -240,7 +240,7 @@ public class UserRatioImage {
                         .drawText(musicInfo.getName(), 160 + dx2, 624 + dy2, new TextEffect().setMaxWidth(220))
                         .font(scoreFont).drawText(String.valueOf(musicInfo.getScore()), 160 + dx2, 646 + dy2)
                         .font(comboMissAccFont)
-                        .drawText("%d\n%d\n%.2f%%".formatted(musicInfo.getCombo(), musicInfo.getMiss(), musicInfo.getAccuracy()), 230 + dx2, 725 + dy2,
+                        .drawText("%d\n%d\n%.2f%%".formatted(musicInfo.getCombo(), musicInfo.getMiss(), (float) musicInfo.getAccuracy()), 230 + dx2, 725 + dy2,
                                 new TextEffect().setSpaceHeight(1))
                         .drawText("> %d (%s)".formatted(musicInfo.getRatioInt(), diff), 163 + dx2, 702 + dy2)
                         .font(levelFont, Color.WHITE)
